Copied files over
I moved the files to be synced to thee GitHub repository. These files include the Processing Java files to run the game in Processing. It also has the exported application files so the game could be run standalone. This exported version was exported today. I have also included a copy of the sound library that I had downloaded through Processing 'Add library' tool to get the sound to work.